Abstracted Navbar Code
 Placed each section of the navbar in it's own component file
